Revert "Refactor functions (#16)"
This reverts commit 70212c9b40c54d056d33a7e7c1b7fe197a97026f.
