+++ functions/package.json
@@ -21,11 +21,8 @@
   "main": "lib/index.js",
   "dependencies": {
     "axios": "^0.27.2",
-    "body-parser": "^1.20.0",
     "cap-ts": "git+https://github.com/philgresh/cap-ts#bc599f9",
-    "cors": "^2.8.5",
     "dotenv": "^16.0.1",
-    "express": "^4.18.1",
     "fast-xml-parser": "^4.0.9",
     "firebase-admin": "^10.0.2",
     "firebase-functions": "^3.23.0",
@@ -40,7 +37,6 @@
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-typescript": "^7.17.12",
     "@firebase/testing": "^0.20.11",
-    "@types/cors": "^2.8.12",
     "@types/jest": "^28.1.8",
     "@types/luxon": "^3.0.1",
     "@types/sinon": "^10.0.13",
